pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "lead-agent"
version = "0.1.0"
description = "AI Lead Agent system with configurable workflows and multi-agent communication"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "lead_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.0"
pyyaml = "^6.0.1"
asyncio = "^3.4.3"
aiohttp = "^3.9.0"
tenacity = "^8.2.3"
structlog = "^23.2.0"
typing-extensions = "^4.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/lead_agent --cov-report=html --cov-report=term-missing --cov-fail-under=0"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
